requirements.txt
streamlit
openai
python-dotenv
google-genai
google-generativeai
python-docx
fitz